+++ demo/server.php
@@ -1,0 +1,6 @@
+<?php
+
+include __DIR__ . '/bootstrap.php';
+
+$fc = $debugbar->createServerHandlerFrontController();
+$fc->handle();