+++ src/DebugBar/Bridge/CacheCache/CacheCacheCollector.php
@@ -8,11 +8,16 @@
  * file that was distributed with this source code.
  */
 
-namespace DebugBar\Bridge;
+namespace DebugBar\Bridge\CacheCache;
 
 use CacheCache\Cache;
 use CacheCache\LoggingBackend;
 use Monolog\Logger;
+use DebugBar\Bridge\MonologCollector;
+use DebugBar\DataCollector\AssetProvider;
+use DebugBar\ServerHandler\ServerHandlerFactoryInterface;
+use DebugBar\Widget\CacheTab;
+use DebugBar\Widget\DataMap;
 
 /**
  * Collects CacheCache operations
@@ -27,9 +32,11 @@ use Monolog\Logger;
  * $debugbar['cache']->addCache(CacheManager::get('default'));
  * </code>
  */
-class CacheCacheCollector extends MonologCollector
+class CacheCacheCollector extends MonologCollector implements AssetProvider, ServerHandlerFactoryInterface
 {
     protected $logger;
+
+    protected $caches = array();
 
     public function __construct(Cache $cache = null, Logger $logger = null, $level = Logger::DEBUG, $bubble = true)
     {
@@ -45,6 +52,11 @@ class CacheCacheCollector extends MonologCollector
         }
     }
 
+    /**
+     * Adds another cache backend which operations should be logged
+     *
+     * @param Cache $cache
+     */
     public function addCache(Cache $cache)
     {
         $backend = $cache->getBackend();
@@ -53,10 +65,34 @@ class CacheCacheCollector extends MonologCollector
         }
         $cache->setBackend($backend);
         $this->addLogger($backend->getLogger());
+        $this->caches[] = $cache;
+        return $this;
     }
 
     public function getName()
     {
         return 'cache';
+    }
+
+    public function getWidgets()
+    {
+        $name = $this->getName();
+        return array(
+            $name => new CacheTab("suitcase", "$name.records"),
+            "$name:badge" => new DataMap("$name.count")
+        );
+    }
+
+    public function getAssets()
+    {
+        return array(
+            'css' => 'widgets/cache/widget.css',
+            'js' => 'widgets/cache/widget.js'
+        );
+    }
+
+    public function getServerHandler()
+    {
+        return new CacheCacheHandler($this->caches);
     }
 }